GLUKIs Prefab made and tested with in GLUKITest scene
I've saved GLUKIs as a prefab and have them in the scene called GLUKITest. At the moment, they just look like white circles that drift toward the center cell and are able to bounce off of walls. As the cell walls at the moment don't distinguish between what's being bounced off them before dying, I removed the BrickBehavior component on some of the walls in GLUKITest to better demo the GLUKIs bouncing.