Created CellCenterBehavior and BloodSugarLevel in the GameManager
GLUKIs are now destroyed when they collide into the CellCenter to mirror glucose being absorbed from the bloodstream into the cell. The foundation for the BSL meter are now in too and are attatched to the GameManager. All values are still tentative for it though (waiting for Staci's confirmation) and there is no UI for it yet.